--- client/src/Address.js
@@ -1,28 +1,0 @@
-import React from 'react';
-
-export default function Address() {
-  return (
-    <ul className="address-field">
-      <li className="li-header">
-        <h2>address</h2>
-      </li>
-      <li>Catalyst Institute for Creative Arts and Technology GmbH</li>
-      <li>Funkhaus Berlin</li>
-      <li>Nalepastraße 18-50</li>
-      <li>12459 Berlin</li>
-      <li className="li-header">
-        <h2>telephone</h2>
-      </li>
-      <li>+49 30 2900 9052</li>
-      <li className="li-header">
-        <h2>email</h2>
-      </li>
-      <li>Admissions - admissions@catalyst-berlin.com</li>
-      <li>Marketing & editorial - storytelling@catalyst-berlin.com</li>
-      <li>
-        Student experience & volunteering opportunities -
-        Hannah@catalyst-berlin.com
-      </li>
-    </ul>
-  );
-}
