+++ client/src/components/CatalystSocials.js
@@ -1,0 +1,32 @@
+import React from 'react';
+
+export function Socials() {
+  return (
+    <div className="social-links">
+      <a
+        href="https://www.instagram.com/etikettradio/"
+        target="_blank"
+        rel="nofollow noreferrer">
+        <span className="fa fa-instagram" />
+      </a>
+      <a
+        href="https://www.facebook.com/etikettradio"
+        target="_blank"
+        rel="nofollow noreferrer">
+        <span className="fa fa-facebook" />
+      </a>
+      <a
+        href="https://soundcloud.com/etikettradio"
+        target="_blank"
+        rel="nofollow noreferrer">
+        <span className="fa fa-soundcloud" />
+      </a>
+      <a
+        href="https://www.twitch.tv/etikett_radio"
+        target="_blank"
+        rel="nofollow noreferrer">
+        <span className="fa fa-twitch" />
+      </a>
+    </div>
+  );
+}
