+++ client/src/header/Header.js
@@ -12,7 +12,7 @@ import Loading from '../loading/Loading';
 import microphone from '../icons/microphone.png';
 import clapperboard from '../icons/clapperboard.png';
 import MessageBar from './MessageBar';
-import { Socials } from "../App";
+import { Socials } from "../components/CatalystSocials";
 
 function Header({ location, name, setName, isMobileWidth, isMobileDevice }) {
   const context = useContext(Context);
