+++ client/src/style/_Global.scss
@@ -182,6 +182,10 @@ footer {
   }
 }
 
+.social-links {
+  text-align: left;
+}
+
 .social-links > a {
   padding: 1rem;
   font-size: 2rem;
